add logs and rename testdata folder
