+++ src/rpc/service_eth_test.go
@@ -5,7 +5,7 @@ import (
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/shutter-network/encrypting-rpc-server/cache"
 	"github.com/shutter-network/encrypting-rpc-server/rpc"
-	testdata "github.com/shutter-network/encrypting-rpc-server/test-data"
+	"github.com/shutter-network/encrypting-rpc-server/testdata"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"math/big"
