+++ src/testdata/test-data.go
@@ -1,4 +1,4 @@
-package test_data
+package testdata
 
 import (
 	"crypto/ecdsa"
